requirements.txt
fastapi
uvicorn
apscheduler
bdshare
python-dotenv
